+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Fixed
+  - Bug in SVF/glTF downloads after switching to another environment
+
 ## [2.6.1] - 2024-04-11
 
 - Fixed